requirements.txt
fastapi
uvicorn
pytesseract
python-multipart
scikit-learn
pandas
joblib
cloudinary
python-dotenv

